- prototype for blockchain

+++ transaction/txout.go
@@ -1,11 +1,11 @@
 package transaction
 
 type TxOut struct {
-	Value    int64
+	Value    int
 	PkScript []byte
 }
 
-func (self TxOut) NewTxOut(value int64, pkScript []byte) *TxOut {
+func (self TxOut) NewTxOut(value int, pkScript []byte) *TxOut {
 	self = TxOut{
 		Value:    value,
 		PkScript: pkScript,
